requirements.txt
# Core LangChain
langchain
langchain-core
langchain-community

# OpenAI
langchain-openai
openai

# Anthropic
langchain-anthropic

# Google Gemini (PaLM)
langchain-google-genai
google-generativeai

# Hugging Face
langchain-huggingface
transformers
huggingface-hub

# Utils
python-dotenv
numpy
scikit-learn
